refactor(runtime): extract dispatch into its own module
execute_tool_raw, execute_tool, ToolExecContext, and current_agent_depth
all move into tool_runner/dispatch.rs. mod.rs re-exports them via
`pub use self::dispatch::{...}` so every external consumer
(librefang_kernel::kernel::mod, the test suite, auto_dream) keeps
resolving the same paths
(`librefang_runtime::tool_runner::execute_tool_raw`, etc.).

dispatch.rs uses `use super::*;` to pull every sibling-module tool
name (tool_a2a_send, tool_file_read, …) into scope without re-listing
each one — the parent module already aggregates them. Everything
that used to be at mod.rs top (`mcp`, `WebToolsContext`, `SkillRegistry`,
`TaintSink`, `ToolDefinition`/`ToolResult`, `normalize_tool_name`,
`Path`, `debug`/`warn`) follows dispatch.rs because those types are
only referenced by the dispatch path; mod.rs is left with just
`Arc` + `KernelHandle` prelude for `require_kernel`.

mod.rs: 7,976 -> 6,489 LOC (-1487). cargo check + cargo clippy --lib
both clean. The 1.5k-line dispatch table now has its own roof; only
the inline test suite (~6k LOC) remains in mod.rs and is the next
candidate for redistribution.

Refs #3710
